Separate parsing functions into own import
