+++ gsa/src/web/pages/alerts/component.js
@@ -25,31 +25,8 @@ import {
   condition_data_fields,
   method_data_fields,
 } from 'gmp/commands/alerts';
-import {
-  CONDITION_TYPE_ALWAYS,
-  CONDITION_TYPE_FILTER_COUNT_AT_LEAST,
-  CONDITION_TYPE_FILTER_COUNT_CHANGED,
-  CONDITION_TYPE_SEVERITY_AT_LEAST,
-  EVENT_TYPE_UPDATED_SECINFO,
-  EVENT_TYPE_NEW_SECINFO,
-  EVENT_TYPE_TASK_RUN_STATUS_CHANGED,
-  EVENT_TYPE_TICKET_RECEIVED,
-  EVENT_TYPE_ASSIGNED_TICKET_CHANGED,
-  EVENT_TYPE_OWNED_TICKET_CHANGED,
-  METHOD_TYPE_ALEMBA_VFIRE,
-  METHOD_TYPE_SCP,
-  METHOD_TYPE_SEND,
-  METHOD_TYPE_SMB,
-  METHOD_TYPE_SNMP,
-  METHOD_TYPE_SYSLOG,
-  METHOD_TYPE_EMAIL,
-  METHOD_TYPE_START_TASK,
-  METHOD_TYPE_HTTP_GET,
-  METHOD_TYPE_SOURCEFIRE,
-  METHOD_TYPE_VERINICE,
-  METHOD_TYPE_TIPPING_POINT,
-} from 'gmp/models/alert';
-import {isDefined, isString} from 'gmp/utils/identity';
+
+import {isDefined} from 'gmp/utils/identity';
 import {selectSaveId} from 'gmp/utils/id';
 import {first} from 'gmp/utils/array';
 import {capitalizeFirstLetter, shorten} from 'gmp/utils/string';
@@ -105,191 +82,12 @@ import AlertDialog, {
 
 import ContentComposerDialog from './contentcomposerdialog';
 
-const fileToBase64 = file => {
-  return new Promise(resolve => {
-    const reader = new FileReader();
-    reader.onload = function (event) {
-      // Remove data url declaration
-      resolve(event.target.result.substring(33));
-    };
-
-    reader.readAsDataURL(file);
-  });
-};
-
-const convertSecInfoEnum = infoType => {
-  switch (infoType) {
-    case 'nvt':
-      return 'NVT';
-    case 'cve':
-      return 'CVE';
-    case 'cpe':
-      return 'CPE';
-    case 'cert_bund_adv':
-      return 'CERT_BUND_ADV';
-    case 'dfn_cert_adv':
-      return 'DFN_CERT_ADV';
-    case 'ovaldef':
-      return 'OVALDEF';
-    default:
-      return null;
-  }
-};
-
-const convertDirectionEnum = directionType => {
-  switch (directionType) {
-    case 'increased':
-      return 'INCREASED';
-    case 'decreased':
-      return 'DECREASED';
-    case 'changed':
-      return 'CHANGED';
-    default:
-      return null;
-  }
-};
-
-const convertFeedEventEnum = feedEvent => {
-  switch (feedEvent) {
-    case 'new':
-      return 'NEW';
-    case 'updated':
-      return 'UPDATED';
-    default:
-      return null;
-  }
-};
-
-const convertTaskStatusEnum = status => {
-  switch (status) {
-    case 'New':
-      return 'NEW';
-    case 'Done':
-      return 'DONE';
-    case 'Requested':
-      return 'REQUESTED';
-    case 'Running':
-      return 'RUNNING';
-    case 'Stop Requested':
-      return 'STOP_REQUESTED';
-    case 'Stopped':
-      return 'STOPPED';
-    default:
-      return null;
-  }
-};
-
-const convertDeltaTypeEnum = deltaType => {
-  switch (deltaType) {
-    case 'None':
-      return 'NONE';
-    case 'Report':
-      return 'REPORT';
-    case 'Previous':
-      return 'PREVIOUS';
-    default:
-      return null;
-  }
-};
-
-const convertDict = async (prefix, data, fields) => {
-  const fieldDict = {};
-  for (const field of fields) {
-    const name = prefix + '_' + field;
-    if (data.hasOwnProperty(name)) {
-      if (field === 'pkcs12') {
-        const base64File = await fileToBase64(data[name]);
-        fieldDict[field] = base64File;
-      } else if (field === 'secinfo_type') {
-        fieldDict[field] = convertSecInfoEnum(data[name]);
-      } else if (field === 'direction') {
-        fieldDict[field] = convertDirectionEnum(data[name]);
-      } else if (field === 'feed_event') {
-        fieldDict[field] = convertFeedEventEnum(data[name]);
-      } else if (field === 'status') {
-        fieldDict[field] = convertTaskStatusEnum(data[name]);
-      } else if (field === 'delta_type') {
-        fieldDict[field] = convertDeltaTypeEnum(data[name]);
-      } else if (field === 'delta_report_id') {
-        fieldDict[field] = data[name] === '' ? null : data[name];
-      } else if (field === 'send_port' || field === 'defense_center_port') {
-        fieldDict[field] = isString(data[name])
-          ? parseInt(data[name])
-          : data[name];
-      } else {
-        fieldDict[field] = data[name];
-      }
-    }
-  }
-  return fieldDict;
-};
-
-const convertConditionEnum = condition => {
-  switch (condition) {
-    case CONDITION_TYPE_ALWAYS:
-      return 'ALWAYS';
-    case CONDITION_TYPE_FILTER_COUNT_AT_LEAST:
-      return 'FILTER_COUNT_AT_LEAST';
-    case CONDITION_TYPE_FILTER_COUNT_CHANGED:
-      return 'FILTER_COUNT_CHANGED';
-    case CONDITION_TYPE_SEVERITY_AT_LEAST:
-      return 'SEVERITY_AT_LEAST';
-    default:
-      return null;
-  }
-};
-
-const convertEventEnum = event => {
-  // Currently does not support all event types. Need updating.
-  switch (event) {
-    case EVENT_TYPE_TASK_RUN_STATUS_CHANGED:
-      return 'TASK_RUN_STATUS_CHANGED';
-    case EVENT_TYPE_UPDATED_SECINFO:
-      return 'UPDATED_SECINFO_ARRIVED';
-    case EVENT_TYPE_NEW_SECINFO:
-      return 'NEW_SECINFO_ARRIVED';
-    case EVENT_TYPE_TICKET_RECEIVED:
-      return null;
-    case EVENT_TYPE_ASSIGNED_TICKET_CHANGED:
-      return null;
-    case EVENT_TYPE_OWNED_TICKET_CHANGED:
-      return null;
-    default:
-      return null;
-  }
-};
-
-const convertMethodEnum = method => {
-  // Currently does not support all method types. Need updating.
-  switch (method) {
-    case METHOD_TYPE_ALEMBA_VFIRE:
-      return null;
-    case METHOD_TYPE_SCP:
-      return 'SCP';
-    case METHOD_TYPE_SEND:
-      return 'SEND';
-    case METHOD_TYPE_SMB:
-      return 'SMB';
-    case METHOD_TYPE_SNMP:
-      return 'SNMP';
-    case METHOD_TYPE_SYSLOG:
-      return 'SYSLOG';
-    case METHOD_TYPE_EMAIL:
-      return 'EMAIL';
-    case METHOD_TYPE_START_TASK:
-      return 'START_TASK';
-    case METHOD_TYPE_HTTP_GET:
-      return 'HTTP_GET';
-    case METHOD_TYPE_SOURCEFIRE:
-      return 'SOURCEFIRE_CONNECTOR';
-    case METHOD_TYPE_VERINICE:
-      return 'VERINICE_CONNECTOR';
-    case METHOD_TYPE_TIPPING_POINT:
-      return 'TippingPoint SMS';
-    default:
-      return null;
-  }
-};
+import {
+  convertConditionEnum,
+  convertDict,
+  convertMethodEnum,
+  convertEventEnum,
+} from './parser';
 
 const select_verinice_report_id = (report_formats, report_id) => {
   if (isDefined(report_id)) {
